+++ src/idunno.AtProto.Types/AtIdentifier.cs
@@ -25,7 +25,7 @@ public abstract class AtIdentifier
 
         if (!TryParse(s, out AtIdentifier? returnValue))
         {
-            throw new ArgumentException("{s} is not a valid AtIdentifier", nameof(s));
+            throw new ArgumentException($"\"{s}\" is not a valid AtIdentifier", nameof(s));
         }
 
         return returnValue;